Updated record processing to only process one at a time
