+++ config/production.yaml
@@ -1,0 +1,6 @@
+---
+PLAINTEXT_VARIABLES:
+    ENVIRONMENT: production
+    NYPL_DATA_API_BASE_URL: https://platform.nypl.org/api/v0.1/
+    SCHEMA_PATH: current-schemas/
+...
