+++ config/qa.yaml
@@ -1,0 +1,6 @@
+---
+PLAINTEXT_VARIABLES:
+    ENVIRONMENT: qa
+    NYPL_DATA_API_BASE_URL: https://qa-platform.nypl.org/api/v0.1/
+    SCHEMA_PATH: current-schemas/
+...
