--- event_sources.json
@@ -1,1 +1,0 @@
-{}
