--- sample/deployment.env.sample
@@ -1,4 +1,0 @@
-NYPL_DATA_API_BASE_URL=XXX
-SCHEMA_PATH=XXX
-SCHEMA_NAME=XXX
-NODE_ENV=development
