--- src/helpers/log_helper.py
@@ -1,13 +1,0 @@
-# log_level_codes = {
-#     "emergency": 0,
-#     "alert": 1,
-#     "critical": 2,
-#     "error": 3,
-#     "warning": 4,
-#     "notice": 5,
-#     "info": 6,
-#     "debug": 7
-# }
-
-# def get_log_level_code(log_level):
-#     return log_level_codes.get(log_level, "n/a")
