+++ test/conftest.py
@@ -4,9 +4,9 @@ import pytest
 
 # Sets OS vars for entire set of tests
 TEST_ENV_VARS = {
+    "ENVIRONMENT": "test",
     "NYPL_DATA_API_BASE_URL": "https://qa-platform.nypl.org/api/v0.1/",
     "SCHEMA_PATH": "current-schemas/",
-    "LOG_LEVEL": "debug",
 }
